+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- Help icons on non-obvious settings with popover descriptions
+
 ## [0.6.2] - 2026-03-24
 
 ### Added